DRTII-799 Add config to set min & max backoff times
