+++ server/src/main/resources/config/common.conf
@@ -23,17 +23,6 @@ health-check {
 enable-statsd = true
 enable-statsd = ${?ENABLE_STATSD}
 
-legacy-flight-data-cutoff = "1970-01-01T00:00"
-legacy-flight-data-cutoff = ${?LEGACY_FLIGHT_DATA_CUTOFF}
-
-migration {
-  query-interval-ms = 100
-  query-interval-ms = ${?MIGRATION_QUERY_INTERVAL_MS}
-}
-
-multi-day-parallelism = 2
-multi-day-parallelism = ${?MULTI_DAY_PARALLELISM}
-
 googleTrackingCode = ""
 googleTrackingCode =  ${?GOOGLE_TRACKING_CODE}
 
@@ -44,12 +33,6 @@ env = ${?ENV}
 
 contact-email = ${?CONTACT_EMAIL}
 ooh-phone = ${?OOH_PHONE}
-
-persistenceBaseDir = "/tmp"
-persistenceBaseDir = ${?PERSISTENCE_BASE_DIR}
-
-passenger_splits_csv_url: ""
-passenger_splits_csv_url: ${?PAX_SPLITS_CSV_URL}
 
 key-cloak {
   url: ${?KEY_CLOAK_URL}
