+++ server/src/main/resources/config/crunch.conf
@@ -3,8 +3,6 @@ crunch {
   recrunch-on-start = ${?RECRUNCH_ON_START}
   refresh-arrivals-on-start = false
   refresh-arrivals-on-start = ${?REFRESH_ARRIVALS_ON_START}
-  stage-throttle-millis = 60000
-  stage-throttle-millis = ${?STAGE_THROTTLE_MILLIS}
   flex-desks = "false"
   flex-desks = ${?FLEX_DESKS}
   manifests {
