+++ server/src/main/resources/config/persistence.conf
@@ -10,4 +10,11 @@ persistence {
     live-portstate: ${?SNAPSHOT_MB_LIVE_PORTSTATE}
     voyage-manifest: ${?SNAPSHOT_MB_VM}
   }
+
+  on-stop-backoff {
+    minimum-seconds: 10
+    minimum-seconds: ${?PERSISTENCE_ON_STOP_BACKOFF_MIN_SECONDS}
+    maximum-seconds: 30
+    maximum-seconds: ${?PERSISTENCE_ON_STOP_BACKOFF_MAX_SECONDS}
+  }
 }
