+++ server/src/main/scala/actors/DrtSystemInterface.scala
@@ -13,6 +13,7 @@ import akka.pattern.ask
 import akka.stream.scaladsl.{Sink, Source, SourceQueueWithComplete}
 import akka.stream.{Materializer, OverflowStrategy, UniqueKillSwitch}
 import akka.util.Timeout
+import com.typesafe.config.ConfigFactory
 import controllers.{Deskstats, PaxFlow, UserRoleProviderLike}
 import drt.chroma.chromafetcher.ChromaFetcher.{ChromaForecastFlight, ChromaLiveFlight}
 import drt.chroma.chromafetcher.{ChromaFetcher, ChromaFlightMarshallers}
@@ -65,17 +66,20 @@ trait DrtSystemInterface extends UserRoleProviderLike {
   val purgeOldLiveSnapshots = false
   val purgeOldForecastSnapshots = true
 
-  val restartOnStop: RestartOnStop = RestartOnStop(1.second, 10 seconds)
 
   val manifestLookupService: ManifestLookupLike
 
-  val config: Configuration
+  val config: Configuration = new Configuration(ConfigFactory.load)
   val airportConfig: AirportConfig
   val params: DrtConfigParameters = DrtConfigParameters(config)
   val journalType: StreamingJournalLike = StreamingJournal.forConfig(config)
 
   val gateWalkTimesProvider: GateOrStandWalkTime = walkTimeMillisProviderFromCsv(params.gateWalkTimesFilePath)
   val standWalkTimesProvider: GateOrStandWalkTime = walkTimeMillisProviderFromCsv(params.standWalkTimesFilePath)
+
+  private val minBackoffSeconds = config.get[Int]("persistence.on-stop-backoff.minimum-seconds")
+  private val maxBackoffSeconds = config.get[Int]("persistence.on-stop-backoff.maximum-seconds")
+  val restartOnStop: RestartOnStop = RestartOnStop(minBackoffSeconds seconds, maxBackoffSeconds seconds)
 
   val alertsActor: ActorRef = restartOnStop.actorOf(Props(new AlertsActor(now)), "alerts-actor")
   val liveBaseArrivalsActor: ActorRef = restartOnStop.actorOf(Props(new LiveBaseArrivalsActor(params.snapshotMegaBytesLiveArrivals, now, expireAfterMillis)), name = "live-base-arrivals-actor")
