+++ server/src/main/scala/actors/ProdDrtSystem.scala
@@ -35,7 +35,7 @@ object PostgresTables extends {
 
 case class SubscribeResponseQueue(subscriber: SourceQueueWithComplete[ManifestsFeedResponse])
 
-case class ProdDrtSystem(config: Configuration, airportConfig: AirportConfig)
+case class ProdDrtSystem(airportConfig: AirportConfig)
                         (implicit val materializer: Materializer,
                          val ec: ExecutionContext,
                          val system: ActorSystem) extends DrtSystemInterface {
