--- server/src/main/scala/services/PassengerSplitRatioProvider.scala
@@ -1,5 +1,0 @@
-package services
-
-trait PassengerSplitRatioProvider {
-  def splitRatioProvider: SplitsProvider.SplitProvider
-}
